Change how sections are supressed from the TOC
The main reason for this change is to use a less intrusive method. The
old method was a bit too low-level in removing addcontentsline from the
sections and so on. However, other packages used this info for things
such as references, so section references could display completely
incorrect numberings at times with the old scheme. This new scheme only
changes what is displayed, which is less intrusive.

This also has the positive side-effect of hiding subsubsections from the
TOC.