Hide appended papers subsubsections as well in TOC
The template only wants to show one entry per appended paper in the table
of contents. It therefore hides their sections and subsections from
appearing there. This fix adds code for subsubsections as well, in the
same way as was done for sections and subsections previously, as we of
course want to hide those as well.